segment_tree.c file moved up one directory and data_structures/segment_tree directory deleted
