+++ data_structures/segment_tree.c
@@ -1,6 +1,77 @@
 #include <string.h>
 #include <stdlib.h>
-#include "segment_tree.h"
+#include <assert.h>
+#include <stdbool.h>
+#include <stdio.h>
+
+#include <stddef.h>
+
+/**
+ * Test function type
+*/
+typedef void (*test_func_t) (void);
+
+struct segment_tree;
+
+/**
+ * Compare function
+ * 
+ * @return >0 number if a > b, 0 if a = b, <0 number if a < b
+*/
+typedef int (*comp_func_t)(const void *a, const void *b);
+
+/**
+ * Operation function
+ * 
+ * @param a first element
+ * @param b second element
+ * @param result pointer to memory where result of operation should be written in
+*/
+typedef void (*oper_func_t)(const void *a, const void *b, void *result);
+
+/**
+ * Initialization of segment tree
+ * 
+ * @param tree segment tree element pointer
+ * @param size size of array
+ * @param elem_size size of element
+ * @param data array of elements
+*/
+void segment_tree_init(struct segment_tree *tree, size_t size, size_t elem_size, void *data,
+                         comp_func_t comp_func, oper_func_t oper_func);
+
+/**
+ * Builds tree from provided data in segment_tree_init
+ * 
+ * @param tree segment tree element pointer
+*/
+void segment_tree_build(struct segment_tree *tree);
+
+/**
+ * Updates the element of given position to new_elem
+ * 
+ * @param tree segment tree element pointer
+ * @param pos position where to update with new_elem
+ * @param new_elem new value for position pos
+*/
+void segment_tree_update(struct segment_tree *tree, size_t pos, void *new_elem);
+
+/**
+ * Making query on given segment [l, r]
+ * 
+ * @param tree segment tree element pointer
+ * @param l left end of the segment
+ * @param r right end of the segment
+ * @param result pointer to which query result data should be written in
+*/
+void segment_tree_query(struct segment_tree *tree, size_t l, size_t r, void *result);
+
+/**
+ * Disposes memory allocated for segment tree
+ * 
+ * @param tree tree of which memory should be disposed
+*/
+void segment_tree_dispose(struct segment_tree *tree);
 
 /**
  * Structure of segment tree
@@ -74,4 +145,18 @@ void segment_tree_query(struct segment_tree *tree, size_t l, size_t r, void *res
 void segment_tree_dispose(struct segment_tree *tree){
     free(tree -> data);
     free(tree -> t);
+}
+
+/**
+ * The template for test function
+*/
+void test(){
+    assert(true);
+}
+
+/**
+ * Main function
+*/
+int main(int argc, char **argv) {
+    return 0;
 }
